RF: move Credential into is own submodule
